+++ src/svn.ts
@@ -3,7 +3,7 @@ import { window } from "vscode";
 import * as cp from "child_process";
 import * as iconv from "iconv-lite";
 import * as jschardet from "jschardet";
-import * as path from 'path';
+import * as path from "path";
 
 interface CpOptions {
   cwd?: string;
@@ -19,7 +19,7 @@ export interface ISvn {
 function parseVersion(raw: string): string {
   const match = raw.match(/(\d+\.\d+\.\d+ \(r\d+\))/);
 
-  if(match && match[0]) {
+  if (match && match[0]) {
     return match[0];
   }
   return raw.split(/[\r\n]+/)[0];
@@ -28,44 +28,57 @@ function parseVersion(raw: string): string {
 function findSpecificSvn(path: string): Promise<ISvn> {
   return new Promise<ISvn>((c, e) => {
     const buffers: Buffer[] = [];
-    const child = cp.spawn(path, ['--version']);
-    child.stdout.on('data', (b: Buffer) => buffers.push(b));
-    child.on('error', cpErrorHandler(e));
-    child.on('exit', code => code ? e(new Error('Not found')) : c({ path, version: parseVersion(Buffer.concat(buffers).toString('utf8').trim()) }));
+    const child = cp.spawn(path, ["--version"]);
+    child.stdout.on("data", (b: Buffer) => buffers.push(b));
+    child.on("error", cpErrorHandler(e));
+    child.on(
+      "exit",
+      code =>
+        code
+          ? e(new Error("Not found"))
+          : c({
+              path,
+              version: parseVersion(
+                Buffer.concat(buffers)
+                  .toString("utf8")
+                  .trim()
+              )
+            })
+    );
   });
 }
 
 function findSvnDarwin(): Promise<ISvn> {
   return new Promise<ISvn>((c, e) => {
-    cp.exec('which svn', (err, svnPathBuffer) => {
+    cp.exec("which svn", (err, svnPathBuffer) => {
       if (err) {
-        return e('svn not found');
+        return e("svn not found");
       }
 
-      const path = svnPathBuffer.toString().replace(/^\s+|\s+$/g, '');
+      const path = svnPathBuffer.toString().replace(/^\s+|\s+$/g, "");
 
       function getVersion(path: string) {
         // make sure svn executes
-        cp.exec('svn --version', (err, stdout) => {
+        cp.exec("svn --version", (err, stdout) => {
           if (err) {
-            return e('svn not found');
+            return e("svn not found");
           }
 
           return c({ path, version: parseVersion(stdout.trim()) });
         });
       }
 
-      if (path !== '/usr/bin/svn') {
+      if (path !== "/usr/bin/svn") {
         return getVersion(path);
       }
 
       // must check if XCode is installed
-      cp.exec('xcode-select -p', (err: any) => {
+      cp.exec("xcode-select -p", (err: any) => {
         if (err && err.code === 2) {
           // svn is not installed, and launching /usr/bin/svn
           // will prompt the user to install it
 
-          return e('svn not found');
+          return e("svn not found");
         }
 
         getVersion(path);
@@ -76,17 +89,17 @@ function findSvnDarwin(): Promise<ISvn> {
 
 function findSystemSvnWin32(base: string): Promise<ISvn> {
   if (!base) {
-    return Promise.reject<ISvn>('Not found');
+    return Promise.reject<ISvn>("Not found");
   }
 
-  return findSpecificSvn(path.join(base, 'TortoiseSVN', 'bin', 'svn.exe'));
+  return findSpecificSvn(path.join(base, "TortoiseSVN", "bin", "svn.exe"));
 }
 
 function findSvnWin32(): Promise<ISvn> {
-  return findSystemSvnWin32(process.env['ProgramW6432'])
-    .then(void 0, () => findSystemSvnWin32(process.env['ProgramFiles(x86)']))
-    .then(void 0, () => findSystemSvnWin32(process.env['ProgramFiles']))
-    .then(void 0, () => findSpecificSvn('svn'));
+  return findSystemSvnWin32(process.env["ProgramW6432"])
+    .then(void 0, () => findSystemSvnWin32(process.env["ProgramFiles(x86)"]))
+    .then(void 0, () => findSystemSvnWin32(process.env["ProgramFiles"]))
+    .then(void 0, () => findSpecificSvn("svn"));
 }
 
 export function findSvn(hint: string | undefined): Promise<ISvn> {
@@ -95,12 +108,15 @@ export function findSvn(hint: string | undefined): Promise<ISvn> {
   return first
     .then(void 0, () => {
       switch (process.platform) {
-        case 'darwin': return findSvnDarwin();
-        case 'win32': return findSvnWin32();
-        default: return findSpecificSvn('svn');
+        case "darwin":
+          return findSvnDarwin();
+        case "win32":
+          return findSvnWin32();
+        default:
+          return findSpecificSvn("svn");
       }
     })
-    .then(null, () => Promise.reject(new Error('Svn installation not found.')));
+    .then(null, () => Promise.reject(new Error("Svn installation not found.")));
 }
 
 function cpErrorHandler(cb: (reason?: any) => void): (reason?: any) => void {
@@ -108,8 +124,8 @@ function cpErrorHandler(cb: (reason?: any) => void): (reason?: any) => void {
     if (/ENOENT/.test(err.message)) {
       err = new SvnError({
         error: err,
-        message: 'Failed to execute svn (ENOENT)',
-        svnErrorCode: 'NotASvnRepository'
+        message: "Failed to execute svn (ENOENT)",
+        svnErrorCode: "NotASvnRepository"
       });
     }
 
@@ -128,7 +144,6 @@ export interface ISvnErrorData {
 }
 
 export class SvnError {
-
   error?: Error;
   message: string;
   stdout?: string;
@@ -145,7 +160,7 @@ export class SvnError {
       this.error = void 0;
     }
 
-    this.message = this.message || data.message || 'SVN error';
+    this.message = this.message || data.message || "SVN error";
     this.stdout = data.stdout;
     this.stderr = data.stderr;
     this.exitCode = data.exitCode;
@@ -154,13 +169,20 @@ export class SvnError {
   }
 
   toString(): string {
-    let result = this.message + ' ' + JSON.stringify({
-      exitCode: this.exitCode,
-      svnErrorCode: this.svnErrorCode,
-      svnCommand: this.svnCommand,
-      stdout: this.stdout,
-      stderr: this.stderr
-    }, null, 2);
+    let result =
+      this.message +
+      " " +
+      JSON.stringify(
+        {
+          exitCode: this.exitCode,
+          svnErrorCode: this.svnErrorCode,
+          svnCommand: this.svnCommand,
+          stdout: this.stdout,
+          stderr: this.stderr
+        },
+        null,
+        2
+      );
 
     if (this.error) {
       result += (<any>this.error).stack;
@@ -247,20 +269,6 @@ export class Svn {
     } catch (error) {
       throw new Error("Unable to find repository root path");
     }
-  }
-
-  public async isSvnAvailable() {
-    return new Promise((resolve, reject) => {
-      cp.exec("svn --version", (error, stdout, stderr) => {
-        if (error) {
-          console.log(stderr);
-          window.showErrorMessage(stderr);
-          reject();
-        }
-
-        resolve();
-      });
-    });
   }
 
   open(repositoryRoot: string, workspaceRoot: string): Repository {